+++ src/resolvers/admin-resolvers.ts
@@ -2,14 +2,14 @@ import Resolver from '@forge/resolver';
 
 import graphqlGateway, { CompassComponentTypeObject } from '@atlassian/forge-graphql';
 import { getGroupProjects } from '../services/fetch-projects';
-import { AuthErrorTypes, GitlabAPIGroup, ResolverResponse, DefaultErrorTypes, FeaturesList } from './resolverTypes';
+import { AuthErrorTypes, GitlabAPIGroup, ResolverResponse, DefaultErrorTypes, FeaturesList } from '../resolverTypes';
 import { connectGroup, InvalidGroupTokenError } from '../services/group';
 
 import { setupAndValidateWebhook } from '../services/webhooks';
 import { disconnectGroup } from '../services/disconnect-group';
 import { getForgeAppId } from '../utils/get-forge-app-id';
 import { getLastSyncTime } from '../services/last-sync-time';
-import { connectedGroupsInfo, getFeatures, groupsAllExisting } from './shared-resolvers';
+import { appId, connectedGroupsInfo, getFeatures, groupsAllExisting } from './shared-resolvers';
 
 const resolver = new Resolver();
 
@@ -90,18 +90,7 @@ resolver.define('features', (): ResolverResponse<FeaturesList> => {
 });
 
 resolver.define('appId', (): ResolverResponse<string> => {
-  try {
-    const forgeAppId = getForgeAppId();
-    return {
-      success: true,
-      data: forgeAppId,
-    };
-  } catch (e) {
-    return {
-      success: false,
-      errors: [{ message: e.message, errorType: DefaultErrorTypes.UNEXPECTED_ERROR }],
-    };
-  }
+  return appId();
 });
 
 export default resolver.getDefinitions();
